pyproject.toml
[project]
name = "zoomcamp-google-fit-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "pandas>=2.2,<2.4",
  "pyarrow",
  "boto3",
  "python-dotenv",
  "dbt-core>=1.10,<1.11",
  "dbt-databricks>=1.10,<1.11",
  "databricks-sql-connector>=4.2.5",
  "streamlit>=1.55.0",
]

